test: reconfigure test setup to work around unstable issues
